an initial attempt at splitting classes into their own files
